requirements.txt
Django==4.1.7
django_phonenumber_field==7.0.2
python-decouple==3.8
